+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- delay runloop stop so websocket closes properly
+
 ---
 
 ## [0.2.1] - 2023-03-23